Introduced scale-in mechanism for feed ingestion.
A feed pipeline may change its shape in terms of number of
operator instances involved in the computational layer.
The processing capacity of the compute operator is evaluated
at runtime. In the scenario where the rate of inflow of data
is less than the combined processing capacity of the compute
operator instances, the pipeline transforms to get rid of excess
compute operator instances. The reverse (scale-out) happens
when the rate of inflow exceeds the combined capacity of
the computational layer of the feed pipeline.